thumbnail quality improvements
